Macovey Task05 First Commit

+++ amcovey/task05/src/main/kotlin/ru/bmstu/iu4/sem05/abstracts/AbstractFigure.kt
@@ -6,5 +6,5 @@ import ru.bmstu.iu4.sem05.interfaces.Printable
 abstract class AbstractFigure(var tag: Int): Drawable, Printable {
     open fun stringify() = "Figure(tag=$tag)"
 
-    abstract fun area(): Double
+    abstract fun area(): Int
 }
